Add PCI_DEVICE_ID_AMD_19H_M60H_ROOT to id_table
This was required for my system (Ryzen 7950x desktop) to load the
`ryzen_smu` device driver.

The device ID was found by looking at `lspci -v` and matching it with
the list in
https://github.com/torvalds/linux/blob/master/arch/x86/kernel/amd_nb.c#L18.
I think we may need to add the other entries in there as well, but I'm
not knowledgeable enough about this to determine that.